reorganize commands, add some todos
